+++ package.json
@@ -68,7 +68,7 @@
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
     "eslint-config-mongodb-js": "^5.0.3",
-    "extract-text-webpack-plugin": "^3.0.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "file-loader": "^0.11.1",
     "font-awesome": "^4.7.0",
     "hadron-app": "^3.2.1",
@@ -77,7 +77,7 @@
     "hadron-react-buttons": "^3.4.0",
     "hadron-react-components": "^3.5.1",
     "hadron-spectron": "^5.0.0",
-    "html-webpack-plugin": "^2.28.0",
+    "html-webpack-plugin": "^3.2.0",
     "ignore-loader": "^0.1.2",
     "istanbul-instrumenter-loader": "^3.0.1",
     "jsdom": "^11.1.0",
@@ -94,7 +94,7 @@
     "less": "^2.7.2",
     "less-loader": "^4.0.5",
     "mocha": "^7.0.1",
-    "mocha-webpack": "^1.1.0",
+    "mocha-webpack": "^2.0.0-beta.0",
     "mongodb-connection-model": "^14.5.1",
     "mongodb-data-service": "^16.5.2",
     "mongodb-js-precommit": "^2.2.1",
@@ -116,11 +116,13 @@
     "sinon": "^8.1.1",
     "sinon-chai": "^3.4.0",
     "style-loader": "^0.18.2",
-    "webpack": "^3.5.5",
-    "webpack-bundle-analyzer": "^2.9.0",
-    "webpack-dev-server": "^2.4.5",
-    "webpack-node-externals": "^1.6.0",
-    "webpack-sources": "^1.0.1",
+    "webpack": "^4.41.5",
+    "webpack-bundle-analyzer": "^3.6.0",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.10.3",
+    "webpack-merge": "^4.2.2",
+    "webpack-node-externals": "^1.7.2",
+    "webpack-sources": "^1.4.3",
     "xvfb-maybe": "^0.2.1"
   },
   "dependencies": {
